chore: fix event attribute and improve comments
