+++ x/core/types/events.go
@@ -44,7 +44,12 @@ const (
 	AttributeCommitment              = "commitment"
 	AttributeDataRequestID           = "dr_id"
 	AttributeDataRequestHeight       = "dr_height"
-	AttributeRevealBody              = "reveal_body"
+	AttributePostedDataRequestHeight = "posted_dr_height"
+	AttributeDataRequestPoster       = "dr_poster"
+	AttributeRevealExitCode          = "reveal_exit_code"
+	AttributeRevealGasUsed           = "reveal_gas_used"
+	AttributeReveal                  = "reveal"
+	AttributeRevealProxyPubKeys      = "reveal_proxy_public_keys"
 	AttributeRevealStdout            = "stdout"
 	AttributeRevealStderr            = "stderr"
 	AttributeDataResultID            = "id"
